cleanup, add test
